+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   CNJR-3503
 - Adds support for optionally prefixing user role_id with "user/" during API key authentication.
   CNJR-5214
+- Added endpoint for getting effective policy
+  CNJR-2040
 
 ### Fixed
 - Dedicated user identifier resolver allowing the user identifiers work like any other resource id. The Conjur internal